renames view.js to grid.js
